requirements.txt
Flask
gunicorn
sklearn
pandas
numpy
PyYAML

-e .